Move xcm configuration to xcm_config.rs
